+++ examples/basic/script.js
@@ -1,4 +1,4 @@
-import { Billing } from '@funnelfox/billing';
+import { PaymentMethod, Billing } from '@funnelfox/billing';
 
 (async function main() {
   let checkout = null;
@@ -27,6 +27,12 @@ import { Billing } from '@funnelfox/billing';
         countryCode: 'US',
       },
       container: '#checkout-container',
+      paymentMethodOrder: [
+        PaymentMethod.PAYMENT_CARD,
+        PaymentMethod.PAYPAL,
+        PaymentMethod.GOOGLE_PAY,
+        PaymentMethod.APPLE_PAY,
+      ],
     });
     createCheckoutButton.disabled = false;
     createCheckoutButton.textContent = 'Create';