+++ src/skins/default/index.ts
@@ -1,12 +1,23 @@
 import template from './template.html';
+import cardTemplate from './card.html';
+import paypalTemplate from './paypal.html';
+import googlePayTemplate from './google-pay.html';
+import applePayTemplate from './apple-pay.html';
 import './styles.css';
 import type { Skin, CardInputElements, SkinFactory } from '../types';
-import type { PaymentMethod } from '../../enums';
+import { PaymentMethod } from '../../enums';
 import {
   CardInputSelectors,
   CheckoutState,
   PrimerWrapperInterface,
 } from '../../types';
+
+const paymentMethodTemplates: Record<PaymentMethod, string> = {
+  [PaymentMethod.PAYMENT_CARD]: cardTemplate,
+  [PaymentMethod.PAYPAL]: paypalTemplate,
+  [PaymentMethod.GOOGLE_PAY]: googlePayTemplate,
+  [PaymentMethod.APPLE_PAY]: applePayTemplate,
+};
 
 class DefaultSkin implements Skin {
   private containerSelector: string;
@@ -14,12 +25,15 @@ class DefaultSkin implements Skin {
   private cardInputElements: CardInputElements;
   private primerWrapper: PrimerWrapperInterface;
   currentPurchaseMethod: PaymentMethod;
+  paymentMethodOrder: PaymentMethod[];
 
   constructor(
     primerWrapper: PrimerWrapperInterface,
-    containerSelector: string
+    containerSelector: string,
+    paymentMethodOrder: PaymentMethod[]
   ) {
     this.containerSelector = containerSelector;
+    this.paymentMethodOrder = paymentMethodOrder;
     const containerEl = document.querySelector<HTMLElement>(containerSelector);
 
     if (!containerEl) {
@@ -62,9 +76,14 @@ class DefaultSkin implements Skin {
       });
     };
 
-    const checkedRadio = Array.from(radioButtons).find(radio => radio.checked);
+    const checkedRadio = Array.from(radioButtons)[0];
+    if (!checkedRadio) {
+      throw new Error(
+        'Default skin accordion initialization error: No radio button found'
+      );
+    }
     setTimeout(() => {
-      handleAccordion(checkedRadio || null);
+      checkedRadio.click();
     }, 0);
 
     radioButtons.forEach(radio => {
@@ -101,6 +120,15 @@ class DefaultSkin implements Skin {
 
   private async init() {
     this.containerEl.insertAdjacentHTML('beforeend', template);
+    const paymentMethodContainers = this.containerEl.querySelector(
+      '#ff-payment-method-containers'
+    );
+    this.paymentMethodOrder.forEach(paymentMethod => {
+      paymentMethodContainers.insertAdjacentHTML(
+        'beforeend',
+        paymentMethodTemplates[paymentMethod]
+      );
+    });
     this.initAccordion();
     this.wireCardInputs();
   }
@@ -256,9 +284,14 @@ class DefaultSkin implements Skin {
 
 const createDefaultSkin: SkinFactory = async (
   primerWrapper: PrimerWrapperInterface,
-  containerSelector: string
+  containerSelector: string,
+  paymentMethodOrder: PaymentMethod[]
 ): Promise<Skin> => {
-  const skin = new DefaultSkin(primerWrapper, containerSelector);
+  const skin = new DefaultSkin(
+    primerWrapper,
+    containerSelector,
+    paymentMethodOrder
+  );
   await skin['init']();
   return skin;
 };